+++ src/main/kotlin/com/tribus/markdown/actions/HeadingDownAction.kt
@@ -1,5 +1,6 @@
 package com.tribus.markdown.actions
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
@@ -7,6 +8,8 @@ import com.intellij.openapi.command.WriteCommandAction
 import com.tribus.markdown.util.MarkdownFileUtil
 
 class HeadingDownAction : AnAction(), MarkdownAction {
+
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
 
     override fun actionPerformed(e: AnActionEvent) {
         val editor = e.getData(CommonDataKeys.EDITOR) ?: return
@@ -36,6 +39,6 @@ class HeadingDownAction : AnAction(), MarkdownAction {
     }
 
     override fun update(e: AnActionEvent) {
-        e.presentation.isEnabledAndVisible = MarkdownFileUtil.isMarkdownFile(e)
+        e.presentation.isEnabled = MarkdownFileUtil.isMarkdownFile(e)
     }
 }