+++ src/main/kotlin/com/tribus/markdown/actions/ToggleTaskListAction.kt
@@ -1,5 +1,6 @@
 package com.tribus.markdown.actions
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
@@ -8,6 +9,8 @@ import com.intellij.openapi.util.TextRange
 import com.tribus.markdown.util.MarkdownFileUtil
 
 class ToggleTaskListAction : AnAction(), MarkdownAction {
+
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
 
     override fun actionPerformed(e: AnActionEvent) {
         val editor = e.getData(CommonDataKeys.EDITOR) ?: return
@@ -40,6 +43,6 @@ class ToggleTaskListAction : AnAction(), MarkdownAction {
     }
 
     override fun update(e: AnActionEvent) {
-        e.presentation.isEnabledAndVisible = MarkdownFileUtil.isMarkdownFile(e)
+        e.presentation.isEnabled = MarkdownFileUtil.isMarkdownFile(e)
     }
 }